Added file of nursing homes in path of Hurricane Harvey
Nursing_Homes_In_Path_Of_Harvey.csv contains a list of
nursing facilities in cities in the path of Hurricane Harvey but *NOT*
under an evacuation order as of the time of this commit.

I am using the following map (the one titled 'How much rainfall
is expected') from the NYT:
https://www.nytimes.com/interactive/2017/08/24/us/hurricane-harvey-texas.html

Using this as a guide, have added nursing homes in Houston, TX to the
list. This ups our number of facilities substantially, from 32 to 102.

I have put the data into the following Google Sheet:
https://docs.google.com/spreadsheets/d/1M0s6mW3SHWslazLE3w4GGLe40ONYbjvLCxibWDhMu2s/edit?usp=sharing

Lat/lon data is available but needs to be copied into the right column
for some entries. Pausing now to avoid fatigue-related copy/paste errors.

Responds to Issue #5 and
https://github.com/benlk/harvey-senior-homes/issues/2